+++ src/main/java/org/jabref/gui/entryeditor/EntryEditor.java
@@ -306,12 +306,12 @@ public class EntryEditor extends BorderPane {
                 preferencesService.getImportFormatPreferences(),
                 fileMonitor,
                 dialogService,
-                stateManager,
                 bibEntryTypesManager,
-                keyBindingRepository);
+                keyBindingRepository,
+                libraryTab.searchQueryProperty());
         tabs.add(sourceTab);
         tabs.add(new LatexCitationsTab(databaseContext, preferencesService, dialogService, directoryMonitorManager));
-        tabs.add(new FulltextSearchResultsTab(stateManager, preferencesService, dialogService, taskExecutor));
+        tabs.add(new FulltextSearchResultsTab(stateManager, preferencesService, dialogService, taskExecutor, libraryTab.searchQueryProperty()));
         tabs.add(new AiSummaryTab(libraryTab.getLibraryTabContainer(), dialogService, preferencesService, aiService, libraryTab.getBibDatabaseContext(), taskExecutor));
         tabs.add(new AiChatTab(libraryTab.getLibraryTabContainer(), dialogService, preferencesService, aiService, libraryTab.getBibDatabaseContext(), taskExecutor));