+++ src/main/java/org/jabref/gui/groups/GroupsPreferences.java
@@ -1,40 +1,77 @@
 package org.jabref.gui.groups;
 
+import java.util.EnumSet;
+
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SetProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleObjectProperty;
+import javafx.beans.property.SimpleSetProperty;
+import javafx.collections.FXCollections;
 
 import org.jabref.model.groups.GroupHierarchyType;
 
+import com.google.common.annotations.VisibleForTesting;
+
 public class GroupsPreferences {
 
-    private final ObjectProperty<GroupViewMode> groupViewMode;
+    private final SetProperty<GroupViewMode> groupViewMode;
     private final BooleanProperty shouldAutoAssignGroup;
     private final BooleanProperty shouldDisplayGroupCount;
     private final ObjectProperty<GroupHierarchyType> defaultHierarchicalContext;
 
-    public GroupsPreferences(GroupViewMode groupViewMode,
+    public GroupsPreferences(boolean viewModeIntersection,
+                             boolean viewModeFilter,
+                             boolean viewModeInvert,
                              boolean shouldAutoAssignGroup,
                              boolean shouldDisplayGroupCount,
                              GroupHierarchyType defaultHierarchicalContext) {
 
-        this.groupViewMode = new SimpleObjectProperty<>(groupViewMode);
+        this.groupViewMode = new SimpleSetProperty<>(FXCollections.observableSet());
+        this.shouldAutoAssignGroup = new SimpleBooleanProperty(shouldAutoAssignGroup);
+        this.shouldDisplayGroupCount = new SimpleBooleanProperty(shouldDisplayGroupCount);
+        this.defaultHierarchicalContext = new SimpleObjectProperty<>(defaultHierarchicalContext);
+
+        if (viewModeIntersection) {
+            this.groupViewMode.add(GroupViewMode.INTERSECTION);
+        }
+        if (viewModeFilter) {
+            this.groupViewMode.add(GroupViewMode.FILTER);
+        }
+        if (viewModeInvert) {
+            this.groupViewMode.add(GroupViewMode.INVERT);
+        }
+    }
+
+    @VisibleForTesting
+    public GroupsPreferences(EnumSet<GroupViewMode> groupViewMode,
+                             boolean shouldAutoAssignGroup,
+                             boolean shouldDisplayGroupCount,
+                             GroupHierarchyType defaultHierarchicalContext) {
+        this.groupViewMode = new SimpleSetProperty<>(FXCollections.observableSet(groupViewMode));
         this.shouldAutoAssignGroup = new SimpleBooleanProperty(shouldAutoAssignGroup);
         this.shouldDisplayGroupCount = new SimpleBooleanProperty(shouldDisplayGroupCount);
         this.defaultHierarchicalContext = new SimpleObjectProperty<>(defaultHierarchicalContext);
     }
 
-    public GroupViewMode getGroupViewMode() {
-        return groupViewMode.getValue();
+    public EnumSet<GroupViewMode> getGroupViewMode() {
+        if (groupViewMode.isEmpty()) {
+            return EnumSet.noneOf(GroupViewMode.class);
+        }
+        return EnumSet.copyOf(groupViewMode);
     }
 
-    public ObjectProperty<GroupViewMode> groupViewModeProperty() {
+    public SetProperty<GroupViewMode> groupViewModeProperty() {
         return groupViewMode;
     }
 
-    public void setGroupViewMode(GroupViewMode groupViewMode) {
-        this.groupViewMode.set(groupViewMode);
+    public void setGroupViewMode(GroupViewMode mode, boolean value) {
+        if (value) {
+            groupViewMode.add(mode);
+        } else {
+            groupViewMode.remove(mode);
+        }
     }
 
     public boolean shouldAutoAssignGroup() {