+++ src/test/java/org/jabref/gui/entryeditor/SourceTabTest.java
@@ -12,6 +12,7 @@ import javafx.stage.Stage;
 import org.jabref.gui.DialogService;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.keyboard.KeyBindingRepository;
+import org.jabref.gui.search.SearchType;
 import org.jabref.gui.undo.CountingUndoManager;
 import org.jabref.gui.util.OptionalObjectProperty;
 import org.jabref.logic.bibtex.FieldPreferences;
@@ -49,8 +50,7 @@ class SourceTabTest {
         area = new CodeArea();
         area.appendText("some example\n text to go here\n across a couple of \n lines....");
         StateManager stateManager = mock(StateManager.class);
-        when(stateManager.activeSearchQueryProperty()).thenReturn(OptionalObjectProperty.empty());
-        when(stateManager.activeGlobalSearchQueryProperty()).thenReturn(OptionalObjectProperty.empty());
+        when(stateManager.activeSearchQuery(SearchType.NORMAL_SEARCH)).thenReturn(OptionalObjectProperty.empty());
         KeyBindingRepository keyBindingRepository = new KeyBindingRepository(Collections.emptyList(), Collections.emptyList());
         ImportFormatPreferences importFormatPreferences = mock(ImportFormatPreferences.class, Answers.RETURNS_DEEP_STUBS);
         when(importFormatPreferences.bibEntryPreferences().getKeywordSeparator()).thenReturn(',');
@@ -64,9 +64,9 @@ class SourceTabTest {
                 importFormatPreferences,
                 new DummyFileUpdateMonitor(),
                 mock(DialogService.class),
-                stateManager,
                 mock(BibEntryTypesManager.class),
-                keyBindingRepository);
+                keyBindingRepository,
+                OptionalObjectProperty.empty());
         pane = new TabPane(
                 new Tab("main area", area),
                 new Tab("other tab", new Label("some text")),